Update search.dart
Constructed a search bar on the top of the page. The search bar allows users to search for other users on the platform.

Uses a CustomScrollView with a SliverToBoxAdapter to position the search bar.
Implemented the search bar using a Container with a Row containing an icon and a TextField.
Styled the search bar with a grey background and a thin border.